Move filterToSQL to separate file
