+++ packages/datastore/datastore/src/storage/adapters/websql/utils.ts
@@ -1,6 +1,5 @@
 import invariant from "tiny-invariant";
 import { generateId } from "../..";
-import { Filter, AllOperators } from "../../..";
 
 // TODO change to constants/enums
 export const prepareStatement = (input: any, type: string = "insert"): [string, any[]] => {
@@ -20,54 +19,6 @@ export const prepareStatement = (input: any, type: string = "insert"): [string, 
         return [statement, vals];
     }
     invariant(false, "Unsupported query type");
-};
-
-type OperatorToSQL = {
-    [P in keyof AllOperators]: (key: string, value: any) => string
-};
-
-const defaultOp = (op: string) => (
-    (key: string, value: any) => {
-        if ("string" === typeof value) {
-            value = `'${value}'`;
-        }
-        return `${key} ${op} ${value}`;
-    }
-);
-
-// TODO contains
-const OperatorToSQLMap: OperatorToSQL = {
-    eq: defaultOp('='),
-    gt: defaultOp('>'),
-    ge: defaultOp('>='),
-    lt: defaultOp('<'),
-    le: defaultOp('<='),
-    ne: defaultOp('!='),
-    in: defaultOp('IN'),
-    contains: defaultOp(''),
-    startsWith: (key, value) => `${key} LIKE '${value}%'`,
-    endsWith: (key, value) => `${key} LIKE '%${value}'`
-};
-
-export const filterToSQL = (filter?: Filter) => {
-    if (!filter) { return ""; };
-
-    const tokens: string[] = [];
-
-    Object.keys(filter).forEach((key) => {
-        if (!(filter[key] instanceof Object)) {
-            tokens.push(`eq=${filter[key]}`);
-            return;
-        }
-        const op = Object.keys(filter[key])[0];
-        const operator = OperatorToSQLMap[(op as keyof AllOperators)];
-        invariant(operator, "Operator not supported");
-
-        const value = filter[key][op];
-        tokens.push(operator(key, value));
-    });
-
-    return `WHERE ${tokens.join(" AND ")} ?`;
 };
 
 export const getType = (type: string): string => {
